Proposal creation page cleanup and simulation feature (#1505)
* Added ability to simulate proposal manually.

* Remove duplicate code in new proposal components.
