+++ packages/i18n/locales/en/translation.json
@@ -169,6 +169,7 @@
     "showInstantiateMessage": "Show Instantiate Message",
     "showQrCode": "Show QR code",
     "showRawData": "Show raw data",
+    "simulate": "Simulate",
     "spend": "Spend",
     "stake": "Stake",
     "stakeAllButProposalDeposit": "Stake all but the {{proposalDeposit}} ${{tokenSymbol}} proposal deposit",
@@ -361,6 +362,7 @@
     "needsVote": "You must select a vote.",
     "nftCollectionNotChosen": "You must create an NFT collection or enter the address of an existing collection before using this action.",
     "nftMetadataNotUploaded": "You must upload the NFT metadata before you can mint.",
+    "noActionsToSimulate": "No actions to simulate.",
     "noCancellableVestingContracts": "There are no vesting contracts that can be cancelled.",
     "noClaimsAvailable": "No claims available.",
     "noCw20Tokens": "There are no CW20 tokens displayed in the treasury.",
@@ -1159,6 +1161,7 @@
     "proposalClosed": "Closed successfully.",
     "proposalCreatedCompleteCompensationCycle": "Proposal published. Complete the compensation cycle by saving the proposal ID.",
     "proposalExecuted": "Executed. Refreshing page...",
+    "proposalSimulation": "Proposal simulated successfully. If executed right now, it would not fail.",
     "ratingsSubmitted": "Ratings submitted.",
     "restaked": "Restaked successfully.",
     "saved": "Saved successfully.",
