+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalMultiple/types.ts
@@ -73,6 +73,10 @@ export interface PublishProposalOptions {
   failedSimulationBypassSeconds?: number
 }
 
+export type SimulateProposal = (
+  newProposalData: NewProposalData
+) => Promise<void>
+
 export type PublishProposal = (
   newProposalData: NewProposalData,
   options?: PublishProposalOptions
@@ -87,6 +91,7 @@ export interface MakeUsePublishProposalOptions {
 }
 
 export type UsePublishProposal = () => {
+  simulateProposal: SimulateProposal
   publishProposal: PublishProposal
   anyoneCanPropose: boolean
   depositUnsatisfied: boolean
