+++ packages/stateless/components/proposal/NewProposal.tsx
@@ -1,78 +1,67 @@
 import {
   Close,
   GavelRounded,
+  Speed,
   Visibility,
   VisibilityOff,
 } from '@mui/icons-material'
 import clsx from 'clsx'
 import Fuse from 'fuse.js'
-import { ComponentType, useCallback, useState } from 'react'
+import { ComponentType, useEffect, useState } from 'react'
 import {
+  FieldValues,
   SubmitErrorHandler,
   SubmitHandler,
-  useFieldArray,
+  UnpackNestedValue,
   useFormContext,
 } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import TimeAgo from 'react-timeago'
-import { v4 as uuidv4 } from 'uuid'
 
 import {
-  ActionCategorySelector,
-  ActionsEditor,
-  ActionsRenderer,
-  Button,
-  FilterableItem,
-  FilterableItemPopup,
-  IconButton,
-  InputErrorMessage,
-  ProposalContentDisplay,
-  RawActionsRenderer,
-  TextAreaInput,
-  TextInput,
-  Tooltip,
-} from '@dao-dao/stateless'
-import {
-  ActionCategoryWithLabel,
   ActiveThreshold,
   BaseNewProposalProps,
-  LoadedAction,
-  LoadedActions,
   LoadingData,
-  StatefulEntityDisplayProps,
-  SuspenseLoaderProps,
 } from '@dao-dao/types'
 import {
-  convertActionsToMessages,
   formatDateTime,
   formatPercentOf100,
   formatTime,
   processError,
-  validateRequired,
 } from '@dao-dao/utils'
 
-import { useWallet, useWalletInfo } from '../../../../../../hooks'
-import { NewProposalData, NewProposalForm } from '../../../types'
+import { Button } from '../buttons'
+import { IconButton } from '../icon_buttons'
+import { FilterableItem, FilterableItemPopup } from '../popup'
+import { Tooltip } from '../tooltip'
 
 enum ProposeSubmitValue {
   Preview = 'Preview',
   Submit = 'Submit',
 }
 
-export interface NewProposalProps
-  extends Pick<
-    BaseNewProposalProps<NewProposalForm>,
-    | 'draft'
-    | 'saveDraft'
-    | 'drafts'
-    | 'loadDraft'
-    | 'unloadDraft'
-    | 'draftSaving'
-    | 'deleteDraft'
-    | 'proposalModuleSelector'
-    | 'actionsReadOnlyMode'
-  > {
-  createProposal: (newProposalData: NewProposalData) => Promise<void>
+type BaseProps<FormData extends FieldValues = any> = Omit<
+  BaseNewProposalProps<FormData>,
+  'onCreateSuccess'
+>
+
+export type NewProposalProps<
+  FormData extends FieldValues = any,
+  ProposalData extends unknown = any
+> = BaseProps<FormData> & {
+  content: {
+    Header: ComponentType
+    Main: ComponentType<Pick<BaseProps<FormData>, 'actionsReadOnlyMode'>>
+    Preview: ComponentType
+  }
+  getProposalDataFromFormData: (
+    formData: UnpackNestedValue<FormData>
+  ) => ProposalData
+  createProposal: (newProposalData: ProposalData) => Promise<void>
+  simulateProposal: (newProposalData: ProposalData) => Promise<void>
+  proposalTitle: string
+  isWalletConnecting: boolean
+  additionalSubmitError?: string
   loading: boolean
   isPaused: boolean
   isActive: boolean
@@ -81,15 +70,20 @@ export interface NewProposalProps
   anyoneCanPropose: boolean
   depositUnsatisfied: boolean
   connected: boolean
-  categories: ActionCategoryWithLabel[]
-  loadedActions: LoadedActions
   simulationBypassExpiration?: Date
-  SuspenseLoader: ComponentType<SuspenseLoaderProps>
-  EntityDisplay: ComponentType<StatefulEntityDisplayProps>
 }
 
-export const NewProposal = ({
+export const NewProposal = <
+  FormData extends FieldValues = any,
+  ProposalData extends unknown = any
+>({
+  content: { Header, Main, Preview },
+  getProposalDataFromFormData,
   createProposal,
+  simulateProposal,
+  proposalTitle,
+  isWalletConnecting,
+  additionalSubmitError,
   loading,
   isPaused,
   isActive,
@@ -98,8 +92,6 @@ export const NewProposal = ({
   anyoneCanPropose,
   depositUnsatisfied,
   connected,
-  categories,
-  loadedActions,
   draft,
   saveDraft,
   drafts,
@@ -109,173 +101,84 @@ export const NewProposal = ({
   deleteDraft,
   simulationBypassExpiration,
   proposalModuleSelector,
-  SuspenseLoader,
-  EntityDisplay,
   actionsReadOnlyMode,
-}: NewProposalProps) => {
+}: NewProposalProps<FormData, ProposalData>) => {
   const { t } = useTranslation()
 
-  // Unpack here because we use these at the top level as well as inside of
-  // nested components.
-  const {
-    register,
-    control,
-    handleSubmit,
-    watch,
-    formState: { errors },
-  } = useFormContext<NewProposalForm>()
+  const { handleSubmit } = useFormContext<FormData>()
 
   const [showPreview, setShowPreview] = useState(false)
   const [showSubmitErrorNote, setShowSubmitErrorNote] = useState(false)
   const [submitError, setSubmitError] = useState('')
 
-  const { isWalletConnecting } = useWallet()
-  const { walletAddress = '', walletProfileData } = useWalletInfo()
-
-  const proposalDescription = watch('description')
-  const proposalTitle = watch('title')
-
-  const { append } = useFieldArray({
-    name: 'actionData',
-    control,
-    shouldUnregister: true,
-  })
-
-  const actionData = watch('actionData') || []
-
-  const onSubmitForm: SubmitHandler<NewProposalForm> = useCallback(
-    ({ title, description, actionData }, event) => {
-      setShowSubmitErrorNote(false)
-      setSubmitError('')
-
-      const nativeEvent = event?.nativeEvent as SubmitEvent
-      const submitterValue = (nativeEvent?.submitter as HTMLInputElement)?.value
-
-      // Preview toggled in onClick handler.
-      if (submitterValue === ProposeSubmitValue.Preview) {
-        return
+  const [holdingAltForSimulation, setHoldingAlt] = useState(false)
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === 'Alt') {
+        setHoldingAlt(true)
       }
-
-      let msgs
-      try {
-        msgs = convertActionsToMessages(loadedActions, actionData)
-      } catch (err) {
-        console.error(err)
-        setSubmitError(
-          processError(err, {
-            forceCapture: false,
-          })
-        )
-        return
+    }
+    const handleKeyUp = (event: KeyboardEvent) => {
+      if (event.key === 'Alt') {
+        setHoldingAlt(false)
       }
+    }
 
-      createProposal({
-        title,
-        description,
-        msgs,
-      })
-    },
-    [createProposal, loadedActions]
-  )
+    document.addEventListener('keydown', handleKeyDown)
+    document.addEventListener('keyup', handleKeyUp)
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown)
+      document.removeEventListener('keyup', handleKeyUp)
+    }
+  }, [])
 
-  const onSubmitError: SubmitErrorHandler<NewProposalForm> = useCallback(() => {
+  const onSubmitForm: SubmitHandler<FormData> = (formData, event) => {
+    setShowSubmitErrorNote(false)
+    setSubmitError('')
+
+    const nativeEvent = event?.nativeEvent as SubmitEvent
+    const submitterValue = (nativeEvent?.submitter as HTMLInputElement)?.value
+
+    // Preview toggled in onClick handler.
+    if (submitterValue === ProposeSubmitValue.Preview) {
+      return
+    }
+
+    let data: ProposalData
+    try {
+      data = getProposalDataFromFormData(formData)
+    } catch (err) {
+      console.error(err)
+      setSubmitError(
+        processError(err, {
+          forceCapture: false,
+        })
+      )
+      return
+    }
+
+    if (holdingAltForSimulation) {
+      simulateProposal(data)
+    } else {
+      createProposal(data)
+    }
+  }
+
+  const onSubmitError: SubmitErrorHandler<FormData> = () => {
     setShowSubmitErrorNote(true)
     setSubmitError('')
-  }, [setShowSubmitErrorNote])
+  }
 
   return (
     <form
       className="flex flex-col gap-6"
       onSubmit={handleSubmit(onSubmitForm, onSubmitError)}
     >
-      <div className="rounded-lg bg-background-tertiary">
-        <div className="flex flex-row items-center justify-between gap-6 border-b border-border-secondary py-4 px-6">
-          <p className="primary-text text-text-body">
-            {t('form.proposalsName')}
-          </p>
-
-          <div className="flex grow flex-col">
-            <TextInput
-              error={errors.title}
-              fieldName="title"
-              placeholder={t('form.proposalsNamePlaceholder')}
-              register={register}
-              validation={[validateRequired]}
-            />
-            <InputErrorMessage error={errors.title} />
-          </div>
-        </div>
-        <div className="flex flex-col gap-4 p-6 pt-5">
-          <p className="primary-text text-text-body">
-            {t('form.description')}
-            <span className="text-text-tertiary">
-              {/* eslint-disable-next-line i18next/no-literal-string */}
-              {' – '}
-              {t('info.supportsMarkdownFormat')}
-            </span>
-          </p>
-
-          <div className="flex flex-col">
-            <TextAreaInput
-              error={errors.description}
-              fieldName="description"
-              placeholder={t('form.proposalsDescriptionPlaceholder')}
-              register={register}
-              rows={5}
-              validation={[validateRequired]}
-            />
-            <InputErrorMessage error={errors.description} />
-          </div>
-        </div>
-      </div>
+      <Header />
 
       {!actionsReadOnlyMode && proposalModuleSelector}
 
-      {actionsReadOnlyMode ? (
-        <ActionsRenderer
-          SuspenseLoader={SuspenseLoader}
-          actionData={actionData.flatMap(({ actionKey, data }, index) => {
-            const { category, action } = (
-              actionKey ? loadedActions[actionKey] || {} : {}
-            ) as Partial<LoadedAction>
-
-            return category && action
-              ? {
-                  id: index.toString(),
-                  category,
-                  action,
-                  data,
-                }
-              : []
-          })}
-        />
-      ) : (
-        <ActionsEditor
-          SuspenseLoader={SuspenseLoader}
-          actionDataErrors={errors?.actionData}
-          actionDataFieldName="actionData"
-          categories={categories}
-          className="-mb-2"
-          loadedActions={loadedActions}
-        />
-      )}
-
-      {!actionsReadOnlyMode && (
-        <div className="self-start">
-          <ActionCategorySelector
-            categories={categories}
-            onSelectCategory={({ key }) => {
-              append({
-                // See `CategorizedActionKeyAndData` comment in
-                // `packages/types/actions.ts` for an explanation of why we need
-                // to append with a unique ID.
-                _id: uuidv4(),
-                categoryKey: key,
-              })
-            }}
-          />
-        </div>
-      )}
+      <Main actionsReadOnlyMode={actionsReadOnlyMode} />
 
       <div className="flex flex-col gap-2 border-y border-border-secondary py-6">
         <div className="flex flex-row items-center justify-between gap-6">
@@ -306,7 +209,9 @@ export const NewProposal = ({
 
             <Tooltip
               title={
-                !connected
+                holdingAltForSimulation
+                  ? undefined
+                  : !connected
                   ? t('error.logInToContinue')
                   : depositUnsatisfied
                   ? t('error.notEnoughForDeposit')
@@ -329,42 +234,58 @@ export const NewProposal = ({
                           ? undefined
                           : Number(activeThreshold.absolute_count.count),
                     })
-                  : undefined
+                  : additionalSubmitError
               }
             >
               <Button
                 disabled={
-                  !connected ||
-                  (!anyoneCanPropose && !isMember.loading && !isMember.data) ||
-                  depositUnsatisfied ||
-                  isPaused ||
-                  !isActive
+                  // Only worry about these wallet-specific conditions if not
+                  // simulating.
+                  (!holdingAltForSimulation &&
+                    (!connected ||
+                      (!anyoneCanPropose &&
+                        !isMember.loading &&
+                        !isMember.data) ||
+                      depositUnsatisfied ||
+                      isPaused ||
+                      !isActive)) ||
+                  // If additional error exists, disable button.
+                  !!additionalSubmitError
                 }
                 loading={loading}
                 type="submit"
                 value={ProposeSubmitValue.Submit}
               >
-                <p>
-                  {simulationBypassExpiration ? (
-                    // If bypassing simulation, change button label and show a
-                    // countdown until simulation bypass expires.
-                    <TimeAgo
-                      date={simulationBypassExpiration}
-                      formatter={(value, _, suffix) =>
-                        suffix === 'from now'
-                          ? t('button.publishAnywayWithCountdown', {
-                              secondsRemaining: value,
-                            })
-                          : // In case the countdown expires before the re-render,
-                            // just show the original button label.
-                            t('button.publish')
-                      }
-                    />
-                  ) : (
-                    t('button.publish')
-                  )}
-                </p>
-                <GavelRounded className="!h-4 !w-4" />
+                {holdingAltForSimulation ? (
+                  <>
+                    <p>{t('button.simulate')}</p>
+                    <Speed className="!h-5 !w-5" />
+                  </>
+                ) : (
+                  <>
+                    <p>
+                      {simulationBypassExpiration ? (
+                        // If bypassing simulation, change button label and show
+                        // a countdown until simulation bypass expires.
+                        <TimeAgo
+                          date={simulationBypassExpiration}
+                          formatter={(value, _, suffix) =>
+                            suffix === 'from now'
+                              ? t('button.publishAnywayWithCountdown', {
+                                  secondsRemaining: value,
+                                })
+                              : // In case the countdown expires before the re-render,
+                                // just show the original button label.
+                                t('button.publish')
+                          }
+                        />
+                      ) : (
+                        t('button.publish')
+                      )}
+                    </p>
+                    <GavelRounded className="!h-4 !w-4" />
+                  </>
+                )}
               </Button>
             </Tooltip>
           </div>
@@ -399,26 +320,7 @@ export const NewProposal = ({
 
         {showPreview && (
           <div className="mt-4 rounded-md border border-border-secondary p-6">
-            <ProposalContentDisplay
-              EntityDisplay={EntityDisplay}
-              createdAt={new Date()}
-              creator={{
-                address: walletAddress,
-                name: walletProfileData.loading
-                  ? { loading: true }
-                  : { loading: false, data: walletProfileData.profile.name },
-              }}
-              description={proposalDescription}
-              innerContentDisplay={
-                actionData.length ? (
-                  <RawActionsRenderer
-                    actionData={actionData}
-                    loadedActions={loadedActions}
-                  />
-                ) : undefined
-              }
-              title={proposalTitle}
-            />
+            <Preview />
           </div>
         )}
       </div>
